Adds chapter on hyperparameter optimization with RandomizedSearchCV and GridSearchCV, results and analysis
